requirements.txt
aiohttp==3.12.13
aiortc==1.13.0
rich==14.0.0
python-dotenv==1.1.0
openai==1.82.1
opencv-python==4.11.0.86